auth/requirements.txt
Click==7.0
ecdsa==0.13
Flask==1.0.2
Flask-Cors==3.0.7
future==0.17.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pyasn1==0.4.5
python-jose==3.0.1
rsa==4.0
six==1.12.0
Werkzeug==0.14.1
